refactor(utils): 抽出 scrollToTop 共用函式
